Extract AI settings in a separate file with one tab per AI
